完成基本增删查改

+++ backend/views.py
@@ -1,7 +1,8 @@
 # -*- coding:utf-8 -*-
 # Last-Modified date 2017/3/4
 from django.shortcuts import render
-from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseRedirect
+from django.core.urlresolvers import reverse
 from django.core.exceptions import ObjectDoesNotExist
 from backend.models import Article, Category
 from backend import forms
@@ -16,9 +17,48 @@ def create_article(request):
     :param request:
     :return: The page for user to typing his/her own article.
     """
-    if request.method == "GET":
-        cate_list = Category.objects.all()
-        return render(request, "create.html", {"categories": cate_list})
+    if request.method == "POST":
+        title = request.POST.get("title")
+        author = request.POST.get("author")
+        categoryID = request.POST.get("category")
+        topped = request.POST.get("topped")
+        status = request.POST.get("status")
+        abstract = request.POST.get("abstract")
+        content = request.POST.get("content")
+        categoryObj = Category.objects.get(pk=categoryID)
+
+        if topped is None:
+            topped = False
+        else:
+            topped = True
+
+        if status is None:
+            status = False
+        else:
+            status = True
+        Article.objects.create(title=title,
+                               author=author,
+                               status=status,
+                               topped=topped,
+                               abstract=abstract,
+                               content=content,
+                               category=categoryObj
+                               )
+        return HttpResponseRedirect(reverse('backend_list_article'))
+
+    else:
+        cateList = Category.objects.all()
+        return render(request, "create.html", {"categories": cateList})
+
+
+def list_article(request):
+    """
+    list all article information from article tables;
+    :param request:
+    :return:
+    """
+    articleList = Article.objects.all()
+    return render(request, 'list.html', {'item_list': articleList})
 
 
 def modified_article(request):
@@ -27,6 +67,43 @@ def modified_article(request):
     :param request:
     :return: The page show the specific article to waiting for user for changes it.
     """
+    if request.method == "POST":
+        title = request.POST.get("title")
+        author = request.POST.get("author")
+        categoryID = request.POST.get("category")
+        topped = request.POST.get("topped")
+        status = request.POST.get("status")
+        abstract = request.POST.get("abstract")
+        content = request.POST.get("content")
+        categoryObj = Category.objects.get(pk=categoryID)
+
+        if topped is None:
+            topped = False
+        else:
+            topped = True
+
+        if status is None:
+            status = False
+        else:
+            status = True
+
+        articlePK = int(request.POST.get("pk"))
+        articleInfo = Article.objects.get(pk=articlePK)
+        articleInfo.title = title
+        articleInfo.author = author
+        articleInfo.category = categoryObj
+        articleInfo.topped = topped
+        articleInfo.status = status
+        articleInfo.abstract = abstract
+        articleInfo.content = content
+        articleInfo.save()
+        return HttpResponseRedirect(reverse('backend_list_article'))
+
+    if request.method == "GET":
+        articlePK = int(request.GET.get("pk"))
+        articleInfo = Article.objects.get(pk=articlePK)
+        cateList = Category.objects.all()
+        return render(request, "modified.html", {"artInfo": articleInfo, "categories": cateList})
 
 
 def delete_article(request):
@@ -35,6 +112,11 @@ def delete_article(request):
     :param request:
     :return: The page of operation result
     """
+    if request.method == "GET":
+        articlePK = int(request.GET.get("pk"))
+        Article.objects.get(pk=articlePK).delete()
+
+    return HttpResponseRedirect(reverse('backend_list_article'))
 
 
 def list_category(request):
@@ -44,7 +126,7 @@ def list_category(request):
     :return:
     """
     category_list = Category.objects.all()
-    return render(request, 'list.html', {'item_list': category_list})
+    return render(request, 'category_list.html', {'item_list': category_list})
 
 
 def create_category(request):
@@ -53,6 +135,16 @@ def create_category(request):
     :param request:
     :return: The page for user to typing his/her own article.
     """
+    if request.method == "POST":
+        nav_status = request.POST.get("nav_status")
+        name = request.POST.get("name")
+        if name != "":
+            if nav_status == "on":
+                nav_status = True
+            else:
+                nav_status = False
+            new_category = Category.objects.create(name=name, nav_status=nav_status)
+        return HttpResponseRedirect(reverse('backend_list_category'))
     return render(request, 'category_create.html')
 
 
@@ -71,7 +163,7 @@ def modified_category(request):
         except ObjectDoesNotExist:
             return HttpResponse("unable to delete, it's not existed")
         modForm = forms.categoryForm(instance=modCategory)
-        return render(request, "modified.html", {"form": modForm, "pk": PK})
+        return render(request, "category_modified.html", {"form": modForm, "pk": PK})
 
     # POST 方法用于提交修改信息。
     elif request.method == "POST":
@@ -92,16 +184,19 @@ def modified_category(request):
             return HttpResponse("unable to delete, it's not existed")
 
 
-def delete_category(request, PK):
+def delete_category(request):
     """
     delete article
     :param request:
     :return: The page of operation result
     """
-    PK = int(PK)
-    try:
-        delCategory = Category.objects.get(pk=PK)
+    if request.method == "GET":
+        PK = int(request.GET.get("id"))
+        try:
+            delCategory = Category.objects.get(pk=PK)
+        except ObjectDoesNotExist:
+            return HttpResponse("unable to delete, it's not existed")
         delCategory.delete()
         return render(request, 'success.html')
-    except ObjectDoesNotExist:
+    else:
         return HttpResponse("unable to delete, it's not existed")
